+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+
+
 # 0.3.0 (2026-09-19)
 
 ## New Features